+++ KnockKnock-iOS/Entity/Home/StoreDetail.swift
@@ -17,3 +17,29 @@ struct StoreDetail: Decodable {
   let locationX: String // 경도
   let locationY: String // 위도
 }
+
+struct StoreDetailDTO: Decodable {
+  let id: Int
+  let name: String
+  let description: String
+  let image: String
+  let shopPromotionNames: [String]
+  let url: String
+  let locationX: String // 경도
+  let locationY: String // 위도
+}
+
+extension StoreDetailDTO {
+  func toDomain() -> StoreDetail {
+    return .init(
+      id: id,
+      name: name,
+      description: description,
+      image: image,
+      shopPromotionNames: shopPromotionNames,
+      url: URL(string: url),
+      locationX: locationX,
+      locationY: locationY
+    )
+  }
+}
